+++ core-ng/src/main/java/core/framework/impl/web/ControllerHolder.java
@@ -9,35 +9,15 @@ import java.lang.reflect.Method;
  */
 public class ControllerHolder {
     final Controller controller;
-    final Method targetMethod;
+    final Method targetMethod;      // targetMethod is used to find associated annotation
     final String controllerInfo;
-
     final boolean skipInterceptor;
     public String action;
 
-    public ControllerHolder(Controller controller) {
-        this(controller, null, false);
-    }
-
-    public ControllerHolder(Controller controller, boolean skipInterceptor) {
-        this(controller, null, skipInterceptor);
-    }
-
-    public ControllerHolder(Controller controller, Method targetMethod) {
-        this(controller, targetMethod, false);
-    }
-
-    private ControllerHolder(Controller controller, Method targetMethod, boolean skipInterceptor) {
+    public ControllerHolder(Controller controller, Method targetMethod, String controllerInfo, boolean skipInterceptor) {
         this.controller = controller;
+        this.targetMethod = targetMethod;
+        this.controllerInfo = controllerInfo;
         this.skipInterceptor = skipInterceptor;
-
-        if (targetMethod == null) {
-            ControllerInspector inspector = new ControllerInspector(controller);
-            this.targetMethod = inspector.targetMethod;
-            controllerInfo = inspector.targetClassName + "." + inspector.targetMethodName;
-        } else {
-            this.targetMethod = targetMethod;
-            controllerInfo = targetMethod.getDeclaringClass().getCanonicalName() + "." + targetMethod.getName();
-        }
     }
 }