+++ core-ng/src/test/java/core/framework/impl/web/ControllerInspectorTest.java
@@ -3,8 +3,12 @@ package core.framework.impl.web;
 import core.framework.api.web.Controller;
 import core.framework.api.web.Request;
 import core.framework.api.web.Response;
-import org.junit.Assert;
 import org.junit.Test;
+
+import static org.hamcrest.CoreMatchers.startsWith;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertThat;
 
 /**
  * @author neo
@@ -13,25 +17,25 @@ public class ControllerInspectorTest {
     @Test
     public void methodReference() throws NoSuchMethodException {
         ControllerInspector inspector = new ControllerInspector(new TestControllers()::get);
-        Assert.assertEquals(TestControllers.class.getCanonicalName(), inspector.targetClassName);
-        Assert.assertEquals("get", inspector.targetMethodName);
-        Assert.assertEquals(TestControllers.class.getDeclaredMethod("get", Request.class), inspector.targetMethod);
+        assertEquals(TestControllers.class, inspector.targetClass);
+        assertEquals(TestControllers.class.getDeclaredMethod("get", Request.class), inspector.targetMethod);
+        assertEquals(TestControllers.class.getCanonicalName() + ".get", inspector.controllerInfo);
     }
 
     @Test
-    public void anonymousMethod() {
+    public void lambdaMethod() {
         ControllerInspector inspector = new ControllerInspector(request -> null);
-        Assert.assertEquals(ControllerInspectorTest.class.getCanonicalName(), inspector.targetClassName);
-        Assert.assertNotNull(inspector.targetMethodName);
-        Assert.assertNotNull(inspector.targetMethod);
+        assertThat(inspector.targetClass.getCanonicalName(), startsWith(ControllerInspectorTest.class.getCanonicalName()));
+        assertNotNull(inspector.targetMethod);
+        assertThat(inspector.controllerInfo, startsWith(ControllerInspectorTest.class.getCanonicalName() + "."));
     }
 
     @Test
     public void staticClass() throws NoSuchMethodException {
         ControllerInspector inspector = new ControllerInspector(new TestController());
-        Assert.assertEquals(TestController.class.getCanonicalName(), inspector.targetClassName);
-        Assert.assertEquals("execute", inspector.targetMethodName);
-        Assert.assertEquals(TestController.class.getMethod("execute", Request.class), inspector.targetMethod);
+        assertEquals(TestController.class, inspector.targetClass);
+        assertEquals(TestController.class.getMethod("execute", Request.class), inspector.targetMethod);
+        assertEquals(TestController.class.getCanonicalName() + ".execute", inspector.controllerInfo);
     }
 
     public static class TestController implements Controller {