+++ core-ng/src/main/java/core/framework/impl/template/parser/HTMLLexer.java
@@ -169,7 +169,11 @@ class HTMLLexer {
     }
 
     private int findAttributeValueLength() {
-        boolean hasDoubleQuote = currentIndex < html.length() && html.charAt(currentIndex) == '"';
+        char currentChar = html.charAt(currentIndex);
+        if (currentChar == '\'')
+            throw Exceptions.error("it is recommended to use double quote to delimit attribute value, location={}", currentLocation());
+
+        boolean hasDoubleQuote = currentIndex < html.length() && currentChar == '"';
         int i = hasDoubleQuote ? currentIndex + 1 : currentIndex;
         for (; i < html.length(); i++) {
             char ch = html.charAt(i);